Develop n

+++ src/main/java/com/base/basesetup/service/AmountInWordsConverterService.java
@@ -1,11 +1,15 @@
 package com.base.basesetup.service;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Service;
 
 @Service
 public interface AmountInWordsConverterService {
 
-	String convert(Long amount);
+	String convert(BigDecimal amount);
+	
+
